requirements.txt
Flask==2.2.2
gunicorn==20.1.0
flask-session==0.4.0
google-auth==2.18.0
google-auth-oauthlib==0.8.0
google-api-python-client==2.70.0
langchain-openai==0.0.1
openai==0.27.0
python-dotenv==0.21.0
requests==2.28.1
oauthlib==3.2.2
